requirements.txt
# TikTok Content Generator — зависимостиopenai>=1.0.0

# Установка: pip install -r requirements.txtgoogle-api-python-client>=2.100.0

yt-dlp>=2024.0.0

# YouTube Data APImoviepy>=2.0.0

google-api-python-client>=2.100.0openai-whisper>=20231117

python-dotenv>=1.0.0

# AI (Groq Cloud — OpenAI-совместимый)requests>=2.31.0

openai>=1.0.0pysrt>=1.1.2

Pillow>=10.0.0

# HTTP-клиент (Telegram bot, Groq proxy)
httpx>=0.27.0

# Транскрипция аудио (faster-whisper быстрее openai-whisper на CPU)
faster-whisper>=1.0.0

# Скачивание аудио с YouTube (с JS-solver для n-challenge)
yt-dlp>=2026.1.1
yt-dlp[default]

# .env файлы
python-dotenv>=1.0.0
